Add a test with counter model
A positive and negative test with history of length 3 was passed. Also,
a conversion function from recorded events to a history of entries was
created. In a real recording, there would be multiple concurrent logs of events
that would need merging and sorting before being converted to a history
of entries.